feat(android): :sparkles: Add alerting feature

+++ src/commonMain/kotlin/shared/SystemInfoData.kt
@@ -3,7 +3,7 @@ package shared
 interface SystemInfoData {
     val cpuUsage: Double
     val memoryUsage: Double
-    val networkRecv: Long
-    val networkSent: Long
+    val networkRecv: Double
+    val networkSent: Double
     val deviceType: String
 }
