+++ src/commonMain/kotlin/shared/SystemMonitor.kt
@@ -4,8 +4,8 @@ interface SystemMonitor {
     val interval: Long
     fun cpuUsage(): Double
     fun memoryUsage(): Double
-    fun networkRecv(): Long
-    fun networkSent(): Long
+    fun networkRecv(): Double
+    fun networkSent(): Double
     fun getDeviceType(): String
     fun run(): SystemInfoData
 }
